refactor(css): mutualize form and input
