+++ src/components/input/GuessingHistory.vue
