+++ src/components/input/GuessingMultipleChoice.vue
